Rewrite gatsby page and node creation to typescript
