+++ .eslintignore
@@ -1,3 +1,2 @@
-/flow-typed/*
 /public/*
 /node_modules/*
