--- gatsby-browser.js
@@ -1,2 +1,0 @@
-require('./src/assets/scss/init.scss');
-require('./static/css/prismjs/theme.min.css');
