+++ gatsby-browser.ts
@@ -1,0 +1,2 @@
+import './src/assets/scss/init.scss';
+import './static/css/prismjs/theme.min.css';
