--- gatsby-node.js
@@ -1,4 +1,0 @@
-require('ts-node').register({ files: true });
-
-exports.createPages = require('./gatsby/create-pages');
-exports.onCreateNode = require('./gatsby/on-create-node');
