+++ gatsby-node.ts
@@ -1,0 +1,2 @@
+export { createPages } from './gatsby/create-pages';
+export { onCreateNode } from './gatsby/on-create-node';
