--- gatsby/create-pages.js
@@ -1,70 +1,0 @@
-const path = require('path');
-const _ = require('lodash');
-const createCategoriesPages = require('./pagination/create-categories-pages');
-const createTagsPages = require('./pagination/create-tags-pages');
-const createPostsPages = require('./pagination/create-posts-pages');
-
-const createPages = async ({ graphql, actions }) => {
-  const { createPage } = actions;
-
-  // 404
-  createPage({
-    path: '/404',
-    component: path.resolve('./src/templates/not-found-template.tsx'),
-  });
-
-  // Tags list
-  createPage({
-    path: '/tags',
-    component: path.resolve('./src/templates/tags-list-template.tsx'),
-  });
-
-  // Categories list
-  createPage({
-    path: '/categories',
-    component: path.resolve('./src/templates/categories-list-template.tsx'),
-  });
-
-  // Posts and pages from markdown
-  const result = await graphql(`
-    {
-      allMarkdownRemark(filter: { frontmatter: { draft: { ne: true } } }) {
-        edges {
-          node {
-            frontmatter {
-              template
-            }
-            fields {
-              slug
-            }
-          }
-        }
-      }
-    }
-  `);
-
-  const { edges } = result.data.allMarkdownRemark;
-
-  edges.forEach(edge => {
-    if (_.get(edge, 'node.frontmatter.template') === 'page') {
-      createPage({
-        path: edge.node.fields.slug,
-        component: path.resolve('./src/templates/page-template.tsx'),
-        context: { slug: edge.node.fields.slug },
-      });
-    } else if (_.get(edge, 'node.frontmatter.template') === 'post') {
-      createPage({
-        path: edge.node.fields.slug,
-        component: path.resolve('./src/templates/post-template.tsx'),
-        context: { slug: edge.node.fields.slug },
-      });
-    }
-  });
-
-  // Feeds
-  await createTagsPages(graphql, actions);
-  await createCategoriesPages(graphql, actions);
-  await createPostsPages(graphql, actions);
-};
-
-module.exports = createPages;
