+++ gatsby/pagination/create-categories-pages.ts
@@ -1,12 +1,17 @@
-const _ = require('lodash');
-const path = require('path');
-const siteConfig = require('../../config.js');
+import { resolve } from 'path';
+import { GatsbyNode } from 'gatsby';
+import { kebabCase } from '../../src/utils/kebabcase';
+import { siteConfig } from '../../config';
+import { AllMarkdownRemark } from '../../src/types';
 
-module.exports = async (graphql, actions) => {
+export const createCategoriesPages: GatsbyNode['createPages'] = async ({
+  graphql,
+  actions,
+}) => {
   const { createPage } = actions;
   const { postsPerPage } = siteConfig;
 
-  const result = await graphql(`
+  const result = await graphql<{ allMarkdownRemark: AllMarkdownRemark }>(`
     {
       allMarkdownRemark(
         filter: { frontmatter: { template: { eq: "post" }, draft: { ne: true } } }
@@ -19,14 +24,22 @@ module.exports = async (graphql, actions) => {
     }
   `);
 
+  if (result.errors) {
+    throw result.errors;
+  }
+
+  if (!result.data) {
+    throw new Error('ERROR: Could not fetch posts on build');
+  }
+
   result.data.allMarkdownRemark.group.forEach(category => {
     const numPages = Math.ceil(category.totalCount / postsPerPage);
-    const categorySlug = `/category/${_.kebabCase(category.fieldValue)}`; // TODO: Get rid of lodash and move kebabcase to tools
+    const categorySlug = `/category/${kebabCase(category.fieldValue)}`;
 
     for (let i = 0; i < numPages; i += 1) {
       createPage({
         path: i === 0 ? categorySlug : `${categorySlug}/page/${i}`,
-        component: path.resolve('./src/templates/category-template.tsx'),
+        component: resolve('./src/templates/category-template.tsx'),
         context: {
           category: category.fieldValue,
           currentPage: i,
