+++ package.json
@@ -9,8 +9,7 @@
     "build": "yarn run clean && gatsby build",
     "deploy": "yarn run clean && gatsby build --prefix-paths && gh-pages -d public",
     "clean": "rimraf .cache public",
-    "lint:js": "eslint --cache --ext .js,.jsx,.ts,.tsx .",
-    "lint": "concurrently \"yarn run lint:js\"",
+    "lint": "eslint --cache --ext .ts,.tsx .",
     "test": "jest --config ./jest/jest.config.js",
     "test:coverage": "jest --coverage --config ./jest/jest.config.js",
     "test:watch": "jest --watch --config ./jest/jest.config.js"
@@ -24,6 +23,9 @@
     "blog",
     "minimal",
     "react"
+  ],
+  "browserslist": [
+    "defaults"
   ],
   "dependencies": {
     "classnames": "^2.2.6",
@@ -51,9 +53,7 @@
     "gatsby-source-filesystem": "^2.1.48",
     "gatsby-transformer-remark": "^2.6.53",
     "gatsby-transformer-sharp": "^2.3.14",
-    "lodash": "^4.17.15",
     "moment": "^2.24.0",
-    "node-sass": "^4.13.1",
     "normalize-scss": "^7.0.1",
     "prismjs": "^1.19.0",
     "react": "^16.12.0",
@@ -67,9 +67,11 @@
     "@babel/plugin-proposal-optional-chaining": "7.8.3",
     "@babel/preset-env": "7.8.4",
     "@babel/preset-react": "7.8.3",
+    "@types/autoprefixer": "^9.7.0",
     "@types/classnames": "^2.2.9",
     "@types/jest": "^25.1.3",
-    "@types/lodash": "^4.14.149",
+    "@types/node": "^13.7.4",
+    "@types/react": "^16.9.22",
     "@types/react-helmet": "^5.0.15",
     "@types/react-test-renderer": "^16.9.2",
     "@typescript-eslint/eslint-plugin": "^2.21.0",
@@ -80,7 +82,6 @@
     "babel-jest": "25.1.0",
     "babel-preset-gatsby": "0.2.29",
     "browserslist": "4.8.7",
-    "concurrently": "5.1.0",
     "eslint": "6.8.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-array-func": "^3.1.3",
@@ -92,12 +93,12 @@
     "eslint-plugin-react": "7.18.3",
     "eslint-plugin-react-hooks": "^2.4.0",
     "eslint-plugin-you-dont-need-lodash-underscore": "^6.9.0",
-    "eslint-watch": "6.0.1",
     "gh-pages": "2.2.0",
     "identity-obj-proxy": "3.0.0",
     "jest": "25.1.0",
     "jest-cli": "25.1.0",
     "lost": "8.3.1",
+    "node-sass": "^4.13.1",
     "postcss-pxtorem": "4.0.1",
     "prettier": "^1.19.1",
     "react-test-renderer": "16.12.0",
