+++ src/globals.d.ts
@@ -3,3 +3,6 @@ declare module '*.scss' {
   const content: { [className: string]: string };
   export = content;
 }
+
+declare module 'lost';
+declare module 'postcss-pxtorem';
