+++ src/templates/tags-list-template.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Link } from 'gatsby';
-import kebabCase from 'lodash/kebabCase';
 import Layout from '../components/Layout';
 import Sidebar from '../components/Sidebar';
 import Page from '../components/Page';
 import useSiteMetadata from '../hooks/use-site-metadata';
 import useTagsList from '../hooks/use-tags-list';
+import { kebabCase } from '../utils/kebabcase';
 
 const TagsListTemplate = () => {
   const { title, subtitle } = useSiteMetadata();
